refactoring(15): extract PaperWorkspace from ResultsPage (last high-risk seam)
Move renderPaper (the Overleaf-like paper editor — file tree, PDF/editor
views, compile log) into PaperWorkspace.tsx. Its state is entangled with
the container's loadResults/loadFiles/openFile data effects, so the state
STAYS in the container (incl. the activeAbsPath mid-body declaration and
its reset effect — the documented hoisting quirk, untouched) and is passed
in as a prop bundle. Pure file helpers (fileIcon/fmtSize/buildFileTree/
isImage/isBinaryPdf/codefileUrl) are module-level; toggleDir/renderTreeNodes/
handleFileClick are nested inside renderPaper verbatim. ResultsPage 922->462
lines (3177 originally, pre-req-03 — an 85% reduction).

Behavior-equivalence adversarially verified (3 skeptics: JSX byte-identity,
file-tree helper-chain fidelity, container wiring incl. activeAbsPath quirk
preservation) — all equivalent, no discrepancies. typecheck 0 non-test
errors; build ok; vitest 4 passed / 2 pre-existing.
